R5-G1-3 (clang-tidy narrowing tier, part 1): 344 benign narrowings made explicit
Adjudicated all 400 bugprone/cppcoreguidelines narrowing-conversions findings
(an 11-agent pass, one per file-group, each classifying every finding REAL vs
BENIGN against the actual code). This commit lands the 344 the review found
provably in-range at the conversion site -- an index or count into a container
whose element count cannot reach the target's limit, a value already clamped or
masked, a loop bound by a small constant. Each is made explicit by wrapping ONLY
the narrowing expression in static_cast to the exact target type (int, or
static_cast<double> for the qint64->double size/duration displays), which is
value-identical: the implicit conversion already happened, this only names it so
clang-tidy's narrowing check stops flagging it and a future reader sees the
truncation is intended and safe.

No value or control flow changed. clang-format + lizard clean, full Release
build + ctest 225/225.

The 55 findings that were NOT auto-cast are handled in the following commits:
15 genuine truncation risks on untrusted parser input (PST/MBOX record and
message counts, email-body HTML offsets) get widened to qsizetype or capped
fail-closed; 40 more qint64->double size displays get an explicit
static_cast<double> (casting them to int would corrupt the value).

+++ src/actions/backup_bitlocker_keys_action.cpp
@@ -689,9 +689,10 @@ bool BackupBitlockerKeysAction::executeExtractKeys(const QDateTime& start_time,
         }
 
         auto& vol = m_volumes[i];
-        const int progress_percent = progress::kStep15 +
-                                     static_cast<int>((static_cast<double>(i) / m_volumes.size()) *
-                                                      kKeyRetrievalProgressSpan);
+        const int progress_percent =
+            progress::kStep15 +
+            static_cast<int>((static_cast<double>(i) / static_cast<double>(m_volumes.size())) *
+                             kKeyRetrievalProgressSpan);
         Q_EMIT executionProgress(QString("Retrieving keys for %1 (%2/%3)...")
                                      .arg(vol.drive_letter)
                                      .arg(i + 1)
@@ -712,7 +713,7 @@ bool BackupBitlockerKeysAction::executeExtractKeys(const QDateTime& start_time,
                              start_time);
             return false;
         }
-        total_keys_found += vol.key_protectors.size();
+        total_keys_found += static_cast<int>(vol.key_protectors.size());
         total_recovery_passwords += countRecoveryPasswords(vol.key_protectors);
 
         // Log protector IDs only (never log the actual recovery passwords)
